requirements.txt
langchain==0.1.12
python-dotenv==1.0.1
streamlit==1.32.0
unstructured==0.12.5
tiktoken==0.6.0
langchain-chroma==0.1.0
langchain-groq==0.1.0
langchain-community==0.0.28
protobuf==4.25.3
langchain-huggingface==0.0.1